fix: retries should not cause tx event handlers to be re-added
I thought about addressing this by having addHandler() be a no-op if the
loop variable tryCnt was not 0, but this might cause weird behavior like if
the tx retries before trying to add its handler.

I like having the post-commit handler defined after the tx logic because
that reflects the order in which they run (otherwise I would've considered
putting it in the options). I suppose we could move options to the end, but
having those up front seems to make more sense as they control how the
(following) tx logic will be executed.

I removed the addHandler() return statement as the function isn't meant to
return anything.

I kept support for multiple event types, but made post-commit the default
to simplify usage for our one and only use case (and perhaps the most common
one).