fix: incrementBy() should generate a condition check if field is read
Noticed this bug when updating the documentation. The old (quantity > 10)
condition _should_ generate a condition check because the increment should
only be processed if the condition check ensures the quantity value that
our increment was conditioned on is still true when the tx commits.

As part of this fix, I removed the prohibition on mixing set() and
incrementBy() calls on NumberField fields.